// src/api/routes/models/AtualizaAppRouteReqDTO.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { AppSubRoute } from './AppSubRoute';

export enum AtualizaAppRouteReqDTOCargosEnum {
  ADMIN = 'ADMIN',
  ESTRUTURA = 'ESTRUTURA',
  PCP = 'PCP',
  USER = 'USER',
  CATERPILLAR_USER = 'CATERPILLAR_USER',
  ALMOX = 'ALMOX',
  LIDER_LASER = 'LIDER_LASER',
  LIDER_SOLDA = 'LIDER_SOLDA',
  LIDER_MONTAGEM = 'LIDER_MONTAGEM',
  LIDER_QUALIDADE = 'LIDER_QUALIDADE',
  LIDER_PROCESSOS = 'LIDER_PROCESSOS',
  DIRETOR = 'DIRETOR',
  LIDER_PCP = 'LIDER_PCP',
  ADMIN_POWERBI = 'ADMIN_POWERBI',
  ADMIN_POWERBI_VIEW = 'ADMIN_POWERBI_VIEW',
}

export type AtualizaAppRouteReqDTO = {
  /**
   * @type array | undefined
   */
  subRoutes?: AppSubRoute[];
  /**
   * @type array | undefined
   */
  cargos?: AtualizaAppRouteReqDTOCargosEnum[];
  /**
   * @type string | undefined
   */
  name?: string;
  /**
   * @type string | undefined
   */
  route?: string;
  /**
   * @type string | undefined
   */
  desc?: string;
};