// src/api/mobile/models/ImpressoraBluetoothResponseDto.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type ImpressoraBluetoothResponseDto = {
  /**
   * @description Identificador único da impressora
   * @type string, uuid
   */
  id: string;
  /**
   * @description Nome da impressora
   * @type string
   */
  nome: string;
  /**
   * @description Endereço MAC da impressora
   * @type string
   */
  macAddress: string;
  /**
   * @description Data de criação do registro
   * @type string, date-time
   */
  dataCriacao: Date;
};